+++ internal/collector/replica_metrics.go
@@ -63,7 +63,7 @@ import (
 	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/inferenceengine"
 	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/logging"
 	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/metrics"
-	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/saturation"
+	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/saturationv1"
 	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/utils"
 	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/utils/scaletarget"
 	llmdVariantAutoscalingV1alpha1 "github.com/llm-d/llm-d-workload-variant-autoscaler/internal/variant"
@@ -932,7 +932,7 @@ func (c *ReplicaMetricsCollector) collectReplicaMetrics(
 		}
 
 		// Look up cost by VariantAutoscaling namespace/name
-		cost := saturation.DefaultVariantCost
+		cost := saturationv1.DefaultVariantCost
 		if variantCosts != nil {
 			if c, ok := variantCosts[variantKey]; ok {
 				cost = c
