+++ internal/saturationv1/analyzer.go
@@ -1,4 +1,4 @@
-package saturation
+package saturationv1
 
 import (
 	"context"
